+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- New tutorial showing how to train an actor using reinforcement learning techniques
+
 ## v1.0.0-beta1 - 2021-04-08
 
 ## v1.0.0-alpha2 - 2021-04-06
 
-### Changed 
+### Changed
 
 - Update to `cogment-py-sdk` v1.0.0-alpha12 and `cogment-orchestrator` v1.0.0-alpha9